+++ apps/server/src/dipeo_server/app_context.py
@@ -34,7 +34,7 @@ def create_server_container() -> Container:
 
     container.registry.register(STATE_STORE, StateRegistry())
 
-    from dipeo.infra import MessageRouter
+    from dipeo.infrastructure.services.messaging import MessageRouter
 
     container.registry.register(MESSAGE_ROUTER, MessageRouter())
