+++ dipeo/application/bootstrap/infrastructure_container.py
@@ -90,7 +90,7 @@ class InfrastructureContainer:
     def _setup_llm_adapter(self):
         """Configure LLM service based on environment."""
         # Use existing LLM infrastructure service
-        from dipeo.infra.llm.service import LLMInfraService
+        from dipeo.infrastructure.services.llm.service import LLMInfraService
         from dipeo.domain.llm import LLMDomainService
 
         # Get API key service from registry
@@ -141,7 +141,7 @@ class InfrastructureContainer:
 
         # Integrated API service
         from dipeo.infrastructure.services.integrated_api import IntegratedApiService
-        from dipeo.infra.adapters.http.api_service import APIService
+        from dipeo.infrastructure.adapters.http.api_service import APIService
         from dipeo.domain.api.services import APIBusinessLogic
         
         # Create API service for providers that need it
