--- dipeo/infra/__init__.py
@@ -1,19 +1,0 @@
-"""DiPeO infrastructure layer.
-
-This module provides concrete implementations of the port interfaces
-defined in the core layer. It includes adapters for external services
-and infrastructure components.
-"""
-
-from .adapters import LLMInfraService, create_adapter
-from .adapters.notion import NotionAPIService
-from .messaging import MessageRouter
-from .persistence.keys import EnvironmentAPIKeyService
-
-__all__ = [
-    "LLMInfraService",
-    "create_adapter",
-    "EnvironmentAPIKeyService",
-    "MessageRouter",
-    "NotionAPIService",
-]
