--- dipeo/infra/adapters/__init__.py
@@ -1,8 +1,0 @@
-"""Infrastructure adapters for external services."""
-
-from ..llm import LLMInfraService, create_adapter
-
-__all__ = [
-    "LLMInfraService",
-    "create_adapter",
-]
