--- dipeo/infra/adapters/notion/__init__.py
@@ -1,5 +1,0 @@
-"""Notion API adapters."""
-
-from .service import NotionAPIService
-
-__all__ = ["NotionAPIService"]
