--- dipeo/infra/persistence/__init__.py
@@ -1,3 +1,0 @@
-"""Infrastructure persistence implementations."""
-
-__all__ = []
