--- dipeo/infra/persistence/keys/__init__.py
@@ -1,5 +1,0 @@
-"""Key management persistence module."""
-
-from dipeo.infra.persistence.keys.environment_service import EnvironmentAPIKeyService
-
-__all__ = ["EnvironmentAPIKeyService"]
