--- dipeo/infra/persistence/memory/__init__.py
@@ -1,5 +1,0 @@
-"""Memory persistence adapters."""
-
-# This module is deprecated - conversation management is now handled by ConversationManagerImpl
-
-__all__ = []
