+++ dipeo/infrastructure/services/llm/factory.py
@@ -1,6 +1,6 @@
 """Factory for creating LLM adapters."""
 
-from .adapters import ChatGPTAdapter, ClaudeAdapter, GeminiAdapter
+from ...adapters.llm import ChatGPTAdapter, ClaudeAdapter, GeminiAdapter
 from .base import BaseLLMAdapter
